pyproject.toml
[project]
name = "askmeanything"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "schroneko", email = "schroneko.fuller@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "discord-py>=2.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/askmeanything"]
